move DeferredWidget to new file, update copyright on routes.dart files
