+++ lib/studies/fortnightly/routes.dart
@@ -1,1 +1,5 @@
+// Copyright 2019 The Flutter team. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 const String defaultRoute = '/fortnightly';
