+++ lib/studies/reply/routes.dart
@@ -1,2 +1,6 @@
+// Copyright 2019 The Flutter team. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 const String homeRoute = '/reply';
 const String composeRoute = '/reply/compose';
